Refactor line-break rewriters
