+++ Reihitsu.Formatter/Pipeline/LineBreaks/LineBreakPhase.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.CodeAnalysis;
+using System.Collections.Generic;
+
+using Microsoft.CodeAnalysis;
 
 namespace Reihitsu.Formatter.Pipeline.LineBreaks;
 
@@ -19,11 +21,40 @@ internal static class LineBreakPhase
     /// <param name="context">The formatting context</param>
     /// <param name="cancellationToken">Cancellation token</param>
     /// <returns>The formatted syntax node with corrected line breaks</returns>
-    public static SyntaxNode Execute(SyntaxNode root, FormattingContext context, CancellationToken cancellationToken)
+    public static SyntaxNode Execute(SyntaxNode root,
+                                     FormattingContext context,
+                                     CancellationToken cancellationToken)
     {
-        var rewriter = new LineBreakRewriter(context, cancellationToken);
+        var current = root;
 
-        return rewriter.Visit(root);
+        foreach (var rewriter in CreateRewriters(context, cancellationToken))
+        {
+            current = rewriter.Visit(current);
+
+            cancellationToken.ThrowIfCancellationRequested();
+        }
+
+        return current;
+    }
+
+    /// <summary>
+    /// Creates the ordered line-break subphase rewriters
+    /// </summary>
+    /// <param name="context">The formatting context</param>
+    /// <param name="cancellationToken">Cancellation token</param>
+    /// <returns>The ordered list of rewriters to execute</returns>
+    private static IReadOnlyList<LineBreakRewriter> CreateRewriters(FormattingContext context,
+                                                                    CancellationToken cancellationToken)
+    {
+        return [
+                   new LineBreakBlockRewriter(context, cancellationToken),
+                   new LineBreakInitializerRewriter(context, cancellationToken),
+                   new LineBreakContainedBlockRewriter(context, cancellationToken),
+                   new LineBreakAssignmentRewriter(context, cancellationToken),
+                   new LineBreakListRewriter(context, cancellationToken),
+                   new LineBreakDeclarationRewriter(context, cancellationToken),
+                   new LineBreakExpressionRewriter(context, cancellationToken),
+               ];
     }
 
     #endregion // Methods
