+++ Reihitsu.Formatter/Pipeline/LineBreaks/LineBreakRewriter.cs
@@ -7,25 +7,11 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 namespace Reihitsu.Formatter.Pipeline.LineBreaks;
 
 /// <summary>
-/// Syntax rewriter that adds and removes line breaks in a single pass.
+/// Shared base syntax rewriter for focused line-break subphases.
 /// Only manipulates <see cref="SyntaxKind.EndOfLineTrivia"/>; does not set indentation
 /// </summary>
-internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
+internal abstract class LineBreakRewriter : CSharpSyntaxRewriter
 {
-    #region Fields
-
-    /// <summary>
-    /// The formatting context
-    /// </summary>
-    private readonly FormattingContext _context;
-
-    /// <summary>
-    /// Cancellation token
-    /// </summary>
-    private readonly CancellationToken _cancellationToken;
-
-    #endregion // Fields
-
     #region Constructor
 
     /// <summary>
@@ -33,42 +19,30 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// </summary>
     /// <param name="context">The formatting context</param>
     /// <param name="cancellationToken">Cancellation token</param>
-    public LineBreakRewriter(FormattingContext context, CancellationToken cancellationToken)
+    protected LineBreakRewriter(FormattingContext context,
+                                CancellationToken cancellationToken)
     {
-        _context = context;
-        _cancellationToken = cancellationToken;
+        Context = context;
+        CancellationToken = cancellationToken;
     }
 
     #endregion // Constructor
 
-    #region Methods
+    #region Properties
 
     /// <summary>
-    /// Collapses the first <see cref="MemberBindingExpressionSyntax"/> in the
-    /// <see cref="ConditionalAccessExpressionSyntax.WhenNotNull"/> subtree onto the
-    /// same line as the <c>?</c> operator token, so that <c>?\n.Member()</c> becomes <c>?.Member()</c>
+    /// Gets the formatting context
     /// </summary>
-    /// <param name="node">The conditional access expression to process</param>
-    /// <returns>The modified node with the member binding collapsed</returns>
-    private static ConditionalAccessExpressionSyntax CollapseMemberBindingToQuestionToken(ConditionalAccessExpressionSyntax node)
-    {
-        var memberBinding = node.WhenNotNull
-                                .DescendantNodesAndSelf()
-                                .OfType<MemberBindingExpressionSyntax>()
-                                .FirstOrDefault();
+    protected FormattingContext Context { get; }
 
-        if (memberBinding == null)
-        {
-            return node;
-        }
+    /// <summary>
+    /// Gets the cancellation token
+    /// </summary>
+    protected CancellationToken CancellationToken { get; }
 
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(memberBinding.OperatorToken) == false)
-        {
-            return node;
-        }
+    #endregion // Properties
 
-        return CollapseTokenToSameLine(node, memberBinding.OperatorToken);
-    }
+    #region Methods
 
     /// <summary>
     /// Determines whether an accessor list belongs to an auto-property
@@ -76,7 +50,7 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// </summary>
     /// <param name="accessorList">The accessor list to inspect</param>
     /// <returns><see langword="true"/> if the accessor list is part of an auto-property; otherwise, <see langword="false"/></returns>
-    private static bool IsAutoPropertyAccessorList(AccessorListSyntax accessorList)
+    protected static bool IsAutoPropertyAccessorList(AccessorListSyntax accessorList)
     {
         foreach (var accessor in accessorList.Accessors)
         {
@@ -90,541 +64,6 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     }
 
     /// <summary>
-    /// Collapses the first argument to the same line as the opening parenthesis
-    /// if it is currently on a new line
-    /// </summary>
-    /// <param name="node">The argument list node</param>
-    /// <returns>The argument list with the first argument collapsed</returns>
-    private static ArgumentListSyntax CollapseFirstArgumentToSameLine(ArgumentListSyntax node)
-    {
-        if (node.Arguments.Count == 0)
-        {
-            return node;
-        }
-
-        var firstArgument = node.Arguments[0];
-        var firstToken = firstArgument.GetFirstToken();
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstToken) == false)
-        {
-            return node;
-        }
-
-        return CollapseTokenToSameLine(node, firstToken);
-    }
-
-    /// <summary>
-    /// Collapses the first argument to the same line as the opening bracket
-    /// if it is currently on a new line
-    /// </summary>
-    /// <param name="node">The bracketed argument list node</param>
-    /// <returns>The argument list with the first argument collapsed</returns>
-    private static BracketedArgumentListSyntax CollapseFirstBracketedArgumentToSameLine(BracketedArgumentListSyntax node)
-    {
-        if (node.Arguments.Count == 0)
-        {
-            return node;
-        }
-
-        var firstArgument = node.Arguments[0];
-        var firstToken = firstArgument.GetFirstToken();
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstToken) == false)
-        {
-            return node;
-        }
-
-        return CollapseTokenToSameLine(node, firstToken);
-    }
-
-    /// <summary>
-    /// Collapses the first attribute argument to the same line as the opening parenthesis
-    /// if it is currently on a new line
-    /// </summary>
-    /// <param name="node">The attribute argument list node</param>
-    /// <returns>The attribute argument list with the first argument collapsed</returns>
-    private static AttributeArgumentListSyntax CollapseFirstAttributeArgumentToSameLine(AttributeArgumentListSyntax node)
-    {
-        if (node.Arguments.Count == 0)
-        {
-            return node;
-        }
-
-        var firstArgument = node.Arguments[0];
-        var firstToken = firstArgument.GetFirstToken();
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstToken) == false)
-        {
-            return node;
-        }
-
-        return CollapseTokenToSameLine(node, firstToken);
-    }
-
-    /// <summary>
-    /// Collapses the first parameter to the same line as the opening parenthesis
-    /// if it is currently on a new line
-    /// </summary>
-    /// <param name="node">The parameter list node</param>
-    /// <returns>The parameter list with the first parameter collapsed</returns>
-    private static ParameterListSyntax CollapseFirstParameterToSameLine(ParameterListSyntax node)
-    {
-        if (node.Parameters.Count == 0)
-        {
-            return node;
-        }
-
-        var firstParameter = node.Parameters[0];
-        var firstToken = firstParameter.GetFirstToken();
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstToken) == false)
-        {
-            return node;
-        }
-
-        return CollapseTokenToSameLine(node, firstToken);
-    }
-
-    /// <summary>
-    /// Collapses the opening parenthesis of a parameter list onto the declaration line
-    /// </summary>
-    /// <param name="node">The parameter list node</param>
-    /// <returns>The updated parameter list</returns>
-    private static ParameterListSyntax CollapseOpenParenToDeclarationLine(ParameterListSyntax node)
-    {
-        if (TryGetPreviousToken(node, node.OpenParenToken, out var previousToken) == false
-            || TokenGapUtilities.HasLineBreakBetween(previousToken, node.OpenParenToken) == false)
-        {
-            return node;
-        }
-
-        var newPreviousToken = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingWhitespace(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia)));
-        var newOpenParen = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(node.OpenParenToken);
-
-        if (ContainsToken(node, previousToken) == false)
-        {
-            return node.WithOpenParenToken(newOpenParen);
-        }
-
-        return node.ReplaceTokens([previousToken, node.OpenParenToken],
-                                  (original, _) => original == previousToken
-                                                       ? newPreviousToken
-                                                       : newOpenParen);
-    }
-
-    /// <summary>
-    /// Collapses misplaced commas onto the previous parameter line
-    /// </summary>
-    /// <param name="node">The parameter list node</param>
-    /// <param name="endOfLine">The end-of-line sequence to keep after a moved separator</param>
-    /// <returns>The updated parameter list</returns>
-    private static ParameterListSyntax CollapseSeparatorsToPreviousParameterLine(ParameterListSyntax node, string endOfLine)
-    {
-        for (var separatorIndex = 0; separatorIndex < node.Parameters.SeparatorCount; separatorIndex++)
-        {
-            var previousToken = node.Parameters[separatorIndex].GetLastToken();
-            var separator = node.Parameters.GetSeparator(separatorIndex);
-            var nextParameter = node.Parameters[separatorIndex + 1].GetFirstToken();
-
-            if (TokenGapUtilities.HasLineBreakBetween(previousToken, separator) == false)
-            {
-                continue;
-            }
-
-            var newPreviousToken = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingWhitespace(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia)));
-            var newSeparator = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(separator);
-
-            if (LineBreakTriviaUtilities.HasTrailingEndOfLine(newSeparator) == false && LineBreakTriviaUtilities.HasLeadingEndOfLine(nextParameter) == false)
-            {
-                var newTrailing = newSeparator.TrailingTrivia
-                                              .Where(trivia => trivia.IsKind(SyntaxKind.WhitespaceTrivia) == false)
-                                              .ToList();
-
-                newTrailing.Add(SyntaxFactory.EndOfLine(endOfLine));
-                newSeparator = newSeparator.WithTrailingTrivia(SyntaxFactory.TriviaList(newTrailing));
-            }
-            node = node.ReplaceTokens([previousToken, separator],
-                                      (original, _) => original == previousToken
-                                                           ? newPreviousToken
-                                                           : newSeparator);
-        }
-
-        return node;
-    }
-
-    /// <summary>
-    /// Collapses the closing parenthesis onto the final parameter line
-    /// </summary>
-    /// <param name="node">The parameter list node</param>
-    /// <returns>The updated parameter list</returns>
-    private static ParameterListSyntax CollapseCloseParenToParameterLine(ParameterListSyntax node)
-    {
-        if (TryGetPreviousToken(node, node.CloseParenToken, out var previousToken) == false
-            || TokenGapUtilities.HasLineBreakBetween(previousToken, node.CloseParenToken) == false)
-        {
-            return node;
-        }
-
-        var newPreviousToken = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingWhitespace(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia)));
-        var newCloseParen = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(node.CloseParenToken);
-
-        return node.ReplaceTokens([previousToken, node.CloseParenToken],
-                                  (original, _) => original == previousToken
-                                                       ? newPreviousToken
-                                                       : newCloseParen);
-    }
-
-    /// <summary>
-    /// Ensures that all arguments in a multi-line argument list start on their own line.
-    /// If the argument list spans multiple lines but some arguments share a line,
-    /// line breaks are inserted after each separator that lacks one
-    /// </summary>
-    /// <param name="node">The argument list node</param>
-    /// <param name="endOfLine">The end-of-line sequence to insert when splitting arguments</param>
-    /// <returns>The argument list with arguments on separate lines</returns>
-    private static ArgumentListSyntax EnsureArgumentsOnSeparateLines(ArgumentListSyntax node, string endOfLine)
-    {
-        if (node.Arguments.Count <= 1)
-        {
-            return node;
-        }
-
-        return EnsureSeparatorsHaveEndOfLine(node, node.Arguments, endOfLine, HasMultiLineAnonymousObjectElement(node.Arguments));
-    }
-
-    /// <summary>
-    /// Ensures that all arguments in a multi-line bracketed argument list start on their own line
-    /// </summary>
-    /// <param name="node">The bracketed argument list node</param>
-    /// <param name="endOfLine">The end-of-line sequence to insert when splitting arguments</param>
-    /// <returns>The argument list with arguments on separate lines</returns>
-    private static BracketedArgumentListSyntax EnsureBracketedArgumentsOnSeparateLines(BracketedArgumentListSyntax node, string endOfLine)
-    {
-        if (node.Arguments.Count <= 1)
-        {
-            return node;
-        }
-
-        return EnsureSeparatorsHaveEndOfLine(node, node.Arguments, endOfLine, HasMultiLineAnonymousObjectElement(node.Arguments));
-    }
-
-    /// <summary>
-    /// Ensures that all arguments in a multi-line attribute argument list start on their own line
-    /// </summary>
-    /// <param name="node">The attribute argument list node</param>
-    /// <param name="endOfLine">The end-of-line sequence to insert when splitting arguments</param>
-    /// <returns>The argument list with arguments on separate lines</returns>
-    private static AttributeArgumentListSyntax EnsureAttributeArgumentsOnSeparateLines(AttributeArgumentListSyntax node, string endOfLine)
-    {
-        if (node.Arguments.Count <= 1)
-        {
-            return node;
-        }
-
-        return EnsureSeparatorsHaveEndOfLine(node, node.Arguments, endOfLine, HasMultiLineAnonymousObjectElement(node.Arguments));
-    }
-
-    /// <summary>
-    /// Ensures that all parameters in a multi-line parameter list start on their own line
-    /// </summary>
-    /// <param name="node">The parameter list node</param>
-    /// <param name="endOfLine">The end-of-line sequence to insert when splitting parameters</param>
-    /// <returns>The parameter list with parameters on separate lines</returns>
-    private static ParameterListSyntax EnsureParametersOnSeparateLines(ParameterListSyntax node, string endOfLine)
-    {
-        if (node.Parameters.Count <= 1)
-        {
-            return node;
-        }
-
-        return EnsureSeparatorsHaveEndOfLine(node, node.Parameters, endOfLine);
-    }
-
-    /// <summary>
-    /// Ensures that each separator in a separated syntax list has a trailing end-of-line trivia.
-    /// Separators that already have a trailing end-of-line (or whose following element has a leading
-    /// end-of-line) are left unchanged
-    /// </summary>
-    /// <typeparam name="TNode">The type of the containing syntax node</typeparam>
-    /// <typeparam name="TElement">The type of the elements in the separated list</typeparam>
-    /// <param name="node">The containing syntax node</param>
-    /// <param name="list">The separated syntax list to process</param>
-    /// <param name="endOfLine">The end-of-line sequence to add after separators that need splitting</param>
-    /// <param name="treatAsMultiLine">Whether the list should be treated as multi-line even after earlier collapsing steps</param>
-    /// <returns>The node with updated separators</returns>
-    private static TNode EnsureSeparatorsHaveEndOfLine<TNode, TElement>(TNode node, SeparatedSyntaxList<TElement> list, string endOfLine, bool treatAsMultiLine = false)
-        where TNode : SyntaxNode
-        where TElement : SyntaxNode
-    {
-        if (treatAsMultiLine == false && IsMultiLine(node) == false)
-        {
-            return node;
-        }
-
-        var hasEndOfLine = treatAsMultiLine;
-        var tokensToReplace = new List<SyntaxToken>();
-        var replacementMap = new Dictionary<SyntaxToken, SyntaxToken>();
-
-        for (var separatorIndex = 0; separatorIndex < list.SeparatorCount; separatorIndex++)
-        {
-            var separator = list.GetSeparator(separatorIndex);
-            var nextElement = list[separatorIndex + 1];
-            var nextFirstToken = nextElement.GetFirstToken();
-
-            if (LineBreakTriviaUtilities.HasTrailingEndOfLine(separator) || LineBreakTriviaUtilities.HasLeadingEndOfLine(nextFirstToken))
-            {
-                hasEndOfLine = true;
-
-                continue;
-            }
-
-            var newTrailing = separator.TrailingTrivia
-                                       .Where(trivia => trivia.IsKind(SyntaxKind.WhitespaceTrivia) == false)
-                                       .ToList();
-
-            newTrailing.Add(SyntaxFactory.EndOfLine(endOfLine));
-
-            tokensToReplace.Add(separator);
-            replacementMap[separator] = separator.WithTrailingTrivia(SyntaxFactory.TriviaList(newTrailing));
-        }
-
-        if (hasEndOfLine == false || tokensToReplace.Count == 0)
-        {
-            return node;
-        }
-
-        return node.ReplaceTokens(tokensToReplace, (original, _) => replacementMap[original]);
-    }
-
-    /// <summary>
-    /// Collapses a multi-line expression-bodied property to a single line
-    /// </summary>
-    /// <param name="node">The property declaration with an expression body</param>
-    /// <returns>The property declaration collapsed to a single line</returns>
-    private static PropertyDeclarationSyntax CollapseExpressionBodiedProperty(PropertyDeclarationSyntax node)
-    {
-        if (node?.ExpressionBody == null)
-        {
-            return null;
-        }
-
-        var updatedNode = node;
-        var arrowToken = updatedNode.ExpressionBody.ArrowToken;
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(arrowToken) || LineBreakTriviaUtilities.HasTrailingEndOfLine(arrowToken.GetPreviousToken()))
-        {
-            updatedNode = CollapseTokenToSameLine(updatedNode, arrowToken);
-            arrowToken = updatedNode.ExpressionBody.ArrowToken;
-        }
-
-        if (arrowToken.LeadingTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-        {
-            updatedNode = updatedNode.ReplaceToken(arrowToken, arrowToken.WithLeadingTrivia(arrowToken.LeadingTrivia.Add(SyntaxFactory.Space)));
-        }
-
-        var firstExpressionToken = updatedNode.ExpressionBody.Expression.GetFirstToken();
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstExpressionToken) || LineBreakTriviaUtilities.HasTrailingEndOfLine(firstExpressionToken.GetPreviousToken()))
-        {
-            updatedNode = CollapseTokenToSameLine(updatedNode, firstExpressionToken);
-        }
-        arrowToken = updatedNode.ExpressionBody.ArrowToken;
-        firstExpressionToken = updatedNode.ExpressionBody.Expression.GetFirstToken();
-
-        var previousToken = arrowToken.GetPreviousToken();
-        var replacementMap = new Dictionary<SyntaxToken, SyntaxToken>
-                             {
-                                 [arrowToken] = arrowToken.WithLeadingTrivia(SyntaxFactory.Space)
-                                                          .WithTrailingTrivia(SyntaxFactory.Space),
-                                 [firstExpressionToken] = firstExpressionToken.WithLeadingTrivia(SyntaxFactory.TriviaList()),
-                             };
-
-        if (previousToken != default && previousToken.IsKind(SyntaxKind.None) == false)
-        {
-            replacementMap[previousToken] = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingWhitespace(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia)));
-        }
-
-        return updatedNode.ReplaceTokens(replacementMap.Keys, (original, _) => replacementMap[original]);
-    }
-
-    /// <summary>
-    /// Determines whether an invocation expression is the outermost node in a method chain.
-    /// An invocation is outermost if it is not an inner link of a larger chain
-    /// and not nested inside a conditional access expression
-    /// </summary>
-    /// <param name="node">The invocation expression to check</param>
-    /// <returns><see langword="true"/> if the invocation is the outermost chain node; otherwise, <see langword="false"/></returns>
-    private static bool IsOutermostChainInvocation(InvocationExpressionSyntax node)
-    {
-        if (node.Expression is not MemberAccessExpressionSyntax
-            && node.Expression is not MemberBindingExpressionSyntax)
-        {
-            return false;
-        }
-
-        if (node.Parent is MemberAccessExpressionSyntax parentAccess
-            && parentAccess.Parent is InvocationExpressionSyntax)
-        {
-            return false;
-        }
-
-        return IsInsideConditionalAccess(node) == false;
-    }
-
-    /// <summary>
-    /// Collects all chain link dot tokens from a method chain or conditional access chain.
-    /// Only invoked member accesses count as chain links.
-    /// For conditional access, the <c>?</c> operator token is collected (not the binding dot)
-    /// </summary>
-    /// <param name="node">The chain node to walk</param>
-    /// <param name="dots">The list to accumulate dot tokens into</param>
-    private static void CollectChainLinkDots(SyntaxNode node, List<SyntaxToken> dots)
-    {
-        switch (node)
-        {
-            case InvocationExpressionSyntax invocation when invocation.Expression is MemberAccessExpressionSyntax memberAccess:
-                {
-                    if (memberAccess.Expression is InvocationExpressionSyntax innerInvocation)
-                    {
-                        CollectChainLinkDots(innerInvocation, dots);
-                    }
-                    else if (memberAccess.Expression is ConditionalAccessExpressionSyntax innerConditional)
-                    {
-                        CollectChainLinkDots(innerConditional, dots);
-                    }
-                    dots.Add(memberAccess.OperatorToken);
-
-                    break;
-                }
-            case ConditionalAccessExpressionSyntax conditionalAccess:
-                {
-                    if (conditionalAccess.Expression is InvocationExpressionSyntax innerInvocation)
-                    {
-                        CollectChainLinkDots(innerInvocation, dots);
-                    }
-                    else if (conditionalAccess.Expression is ConditionalAccessExpressionSyntax innerConditional)
-                    {
-                        CollectChainLinkDots(innerConditional, dots);
-                    }
-                    dots.Add(conditionalAccess.OperatorToken);
-
-                    CollectWhenNotNullChainDots(conditionalAccess.WhenNotNull, dots);
-
-                    break;
-                }
-        }
-    }
-
-    /// <summary>
-    /// Collects chain link dot tokens from the <c>WhenNotNull</c> part of a conditional access expression.
-    /// Skips <see cref="MemberBindingExpressionSyntax"/> dots since they are represented
-    /// by the <c>?</c> operator token of the parent conditional access
-    /// </summary>
-    /// <param name="node">The WhenNotNull expression to walk</param>
-    /// <param name="dots">The list to accumulate dot tokens into</param>
-    private static void CollectWhenNotNullChainDots(SyntaxNode node, List<SyntaxToken> dots)
-    {
-        if (node is InvocationExpressionSyntax invocation
-            && invocation.Expression is MemberAccessExpressionSyntax memberAccess)
-        {
-            CollectWhenNotNullChainDots(memberAccess.Expression, dots);
-
-            dots.Add(memberAccess.OperatorToken);
-        }
-    }
-
-    /// <summary>
-    /// Determines whether an expression contains an invocation expression.
-    /// Used to decide whether a conditional access chain qualifies for chain normalization
-    /// </summary>
-    /// <param name="expression">The expression to inspect</param>
-    /// <returns><see langword="true"/> if the expression contains an invocation; otherwise, <see langword="false"/></returns>
-    private static bool ContainsInvocation(ExpressionSyntax expression)
-    {
-        if (expression is InvocationExpressionSyntax)
-        {
-            return true;
-        }
-
-        foreach (var child in expression.ChildNodes())
-        {
-            if (child is ExpressionSyntax childExpression && ContainsInvocation(childExpression))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Determines whether a syntax node is nested inside a <see cref="ConditionalAccessExpressionSyntax"/>.
-    /// Walks up the parent chain until a statement or member declaration is found
-    /// </summary>
-    /// <param name="node">The node to check</param>
-    /// <returns><see langword="true"/> if the node is inside a conditional access expression; otherwise, <see langword="false"/></returns>
-    private static bool IsInsideConditionalAccess(SyntaxNode node)
-    {
-        var current = node.Parent;
-
-        while (current != null)
-        {
-            if (current is ConditionalAccessExpressionSyntax)
-            {
-                return true;
-            }
-
-            if (current is StatementSyntax || current is MemberDeclarationSyntax)
-            {
-                return false;
-            }
-            current = current.Parent;
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Determines whether a chain dot token has intermediate member accesses between
-    /// the dot and the chain root. This indicates the chain has property accesses
-    /// (e.g., <c>source.Items.Where(...)</c>) and a single invocation link should not
-    /// be collapsed
-    /// </summary>
-    /// <param name="dotToken">The dot token from a member access expression</param>
-    /// <returns><see langword="true"/> if there are intermediate member accesses; otherwise, <see langword="false"/></returns>
-    private static bool HasIntermediateMemberAccess(SyntaxToken dotToken)
-    {
-        if (dotToken.Parent is MemberAccessExpressionSyntax memberAccess)
-        {
-            return memberAccess.Expression is MemberAccessExpressionSyntax
-                   || memberAccess.Expression is ConditionalAccessExpressionSyntax;
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Recursively collects all <c>?</c> and <c>:</c> tokens from a nested ternary expression tree
-    /// </summary>
-    /// <param name="node">The conditional expression to collect tokens from</param>
-    /// <param name="tokens">The list to populate with operator tokens</param>
-    private static void CollectTernaryOperatorTokens(ConditionalExpressionSyntax node, List<SyntaxToken> tokens)
-    {
-        tokens.Add(node.QuestionToken);
-        tokens.Add(node.ColonToken);
-
-        if (node.WhenTrue is ConditionalExpressionSyntax nestedTrue)
-        {
-            CollectTernaryOperatorTokens(nestedTrue, tokens);
-        }
-
-        if (node.WhenFalse is ConditionalExpressionSyntax nestedFalse)
-        {
-            CollectTernaryOperatorTokens(nestedFalse, tokens);
-        }
-    }
-
-    /// <summary>
     /// Collapses a token to the same line as the previous token by removing any
     /// end-of-line trivia from both the token's leading trivia and the previous
     /// token's trailing trivia
@@ -633,7 +72,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="node">The node containing the token</param>
     /// <param name="token">The token to collapse to the previous line</param>
     /// <returns>The node with the token collapsed to the same line</returns>
-    private static TNode CollapseTokenToSameLine<TNode>(TNode node, SyntaxToken token)
+    protected static TNode CollapseTokenToSameLine<TNode>(TNode node,
+                                                          SyntaxToken token)
         where TNode : SyntaxNode
     {
         var newToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(token);
@@ -663,218 +103,11 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// </summary>
     /// <param name="node">The node to inspect</param>
     /// <returns><see langword="true"/> if the node spans multiple lines; otherwise, <see langword="false"/></returns>
-    private static bool IsMultiLine(SyntaxNode node)
+    protected static bool IsMultiLine(SyntaxNode node)
     {
         var text = node.GetText();
 
         return text.Lines.Count > 1;
-    }
-
-    /// <summary>
-    /// Determines whether any element in a separated list contains a multi-line anonymous object
-    /// </summary>
-    /// <typeparam name="TElement">The element syntax type</typeparam>
-    /// <param name="list">The separated list to inspect</param>
-    /// <returns><see langword="true"/> if any element contains a multi-line anonymous object; otherwise, <see langword="false"/></returns>
-    private static bool HasMultiLineAnonymousObjectElement<TElement>(SeparatedSyntaxList<TElement> list)
-        where TElement : SyntaxNode
-    {
-        foreach (var element in list)
-        {
-            foreach (var anonymousObject in element.DescendantNodesAndSelf().OfType<AnonymousObjectCreationExpressionSyntax>())
-            {
-                if (IsMultiLine(anonymousObject))
-                {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Gets the constraint clauses from a syntax node if it supports them
-    /// </summary>
-    /// <param name="node">The syntax node to inspect</param>
-    /// <returns>The list of constraint clauses, or <see langword="null"/> if not applicable</returns>
-    private static SyntaxList<TypeParameterConstraintClauseSyntax>? GetConstraintClauses(SyntaxNode node)
-    {
-        switch (node)
-        {
-            case ClassDeclarationSyntax classDecl:
-                return classDecl.ConstraintClauses;
-
-            case StructDeclarationSyntax structDecl:
-                return structDecl.ConstraintClauses;
-
-            case InterfaceDeclarationSyntax interfaceDecl:
-                return interfaceDecl.ConstraintClauses;
-
-            case RecordDeclarationSyntax recordDecl:
-                return recordDecl.ConstraintClauses;
-
-            case MethodDeclarationSyntax methodDecl:
-                return methodDecl.ConstraintClauses;
-
-            case DelegateDeclarationSyntax delegateDecl:
-                return delegateDecl.ConstraintClauses;
-
-            case LocalFunctionStatementSyntax localFunc:
-                return localFunc.ConstraintClauses;
-
-            default:
-                return null;
-        }
-    }
-
-    /// <summary>
-    /// Sets the constraint clauses on a syntax node
-    /// </summary>
-    /// <typeparam name="TNode">The syntax node type</typeparam>
-    /// <param name="node">The syntax node to modify</param>
-    /// <param name="constraintClauses">The new constraint clauses</param>
-    /// <returns>The node with updated constraint clauses</returns>
-    private static TNode SetConstraintClauses<TNode>(TNode node, SyntaxList<TypeParameterConstraintClauseSyntax> constraintClauses)
-        where TNode : SyntaxNode
-    {
-        switch (node)
-        {
-            case ClassDeclarationSyntax classDecl:
-                return (TNode)(SyntaxNode)classDecl.WithConstraintClauses(constraintClauses);
-
-            case StructDeclarationSyntax structDecl:
-                return (TNode)(SyntaxNode)structDecl.WithConstraintClauses(constraintClauses);
-
-            case InterfaceDeclarationSyntax interfaceDecl:
-                return (TNode)(SyntaxNode)interfaceDecl.WithConstraintClauses(constraintClauses);
-
-            case RecordDeclarationSyntax recordDecl:
-                return (TNode)(SyntaxNode)recordDecl.WithConstraintClauses(constraintClauses);
-
-            case MethodDeclarationSyntax methodDecl:
-                return (TNode)(SyntaxNode)methodDecl.WithConstraintClauses(constraintClauses);
-
-            case DelegateDeclarationSyntax delegateDecl:
-                return (TNode)(SyntaxNode)delegateDecl.WithConstraintClauses(constraintClauses);
-
-            case LocalFunctionStatementSyntax localFunc:
-                return (TNode)(SyntaxNode)localFunc.WithConstraintClauses(constraintClauses);
-
-            default:
-                return node;
-        }
-    }
-
-    /// <summary>
-    /// Removes trailing whitespace from the token immediately before an initializer's close brace
-    /// when the close brace has been moved to a new line
-    /// </summary>
-    /// <param name="node">The initializer expression</param>
-    /// <returns>The initializer with trailing whitespace cleaned up</returns>
-    private static InitializerExpressionSyntax CleanupTrailingWhitespaceBeforeCloseBrace(InitializerExpressionSyntax node)
-    {
-        var closeBrace = node.CloseBraceToken;
-
-        if (closeBrace.LeadingTrivia.Any(SyntaxKind.EndOfLineTrivia) == false)
-        {
-            return node;
-        }
-
-        var previousToken = closeBrace.GetPreviousToken();
-
-        if (previousToken == default
-            || previousToken.IsKind(SyntaxKind.None)
-            || LineBreakTriviaUtilities.HasTrailingEndOfLine(previousToken)
-            || previousToken.TrailingTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-        {
-            return node;
-        }
-
-        var newTrailing = LineBreakTriviaUtilities.RemoveTrailingWhitespace(previousToken.TrailingTrivia);
-        var newPreviousToken = previousToken.WithTrailingTrivia(newTrailing);
-
-        return node.ReplaceToken(previousToken, newPreviousToken);
-    }
-
-    /// <summary>
-    /// Removes trailing whitespace from the token immediately before an initializer's open brace
-    /// when the brace has been moved to a new line
-    /// </summary>
-    /// <typeparam name="TNode">The parent syntax node type</typeparam>
-    /// <param name="node">The parent node containing the initializer</param>
-    /// <param name="initializer">The initializer expression, or <see langword="null"/></param>
-    /// <returns>The node with trailing whitespace cleaned up</returns>
-    private static TNode CleanupTrailingWhitespaceBeforeInitializerBrace<TNode>(TNode node, InitializerExpressionSyntax initializer)
-        where TNode : SyntaxNode
-    {
-        if (initializer == null)
-        {
-            return node;
-        }
-
-        var openBrace = initializer.OpenBraceToken;
-
-        if (openBrace.LeadingTrivia.Any(SyntaxKind.EndOfLineTrivia) == false)
-        {
-            return node;
-        }
-
-        var previousToken = openBrace.GetPreviousToken();
-
-        if (previousToken == default
-            || previousToken.IsKind(SyntaxKind.None)
-            || LineBreakTriviaUtilities.HasTrailingEndOfLine(previousToken)
-            || previousToken.TrailingTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-        {
-            return node;
-        }
-
-        var newTrailing = LineBreakTriviaUtilities.RemoveTrailingWhitespace(previousToken.TrailingTrivia);
-        var newPreviousToken = previousToken.WithTrailingTrivia(newTrailing);
-
-        return node.ReplaceToken(previousToken, newPreviousToken);
-    }
-
-    /// <summary>
-    /// Normalizes a chain containing a single dot token
-    /// </summary>
-    /// <param name="node">The chain node</param>
-    /// <param name="chainDot">The chain dot token</param>
-    /// <returns>The updated chain node</returns>
-    private static SyntaxNode NormalizeSingleChainDot(SyntaxNode node, SyntaxToken chainDot)
-    {
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(chainDot)
-            && HasIntermediateMemberAccess(chainDot) == false)
-        {
-            return CollapseTokenToSameLine(node, chainDot);
-        }
-
-        return node;
-    }
-
-    /// <summary>
-    /// Collapses the first chain dot onto the root line when it starts on a continuation line
-    /// </summary>
-    /// <param name="firstDot">The first chain dot token</param>
-    /// <param name="replacements">The token replacement map to populate</param>
-    private static void TryCollapseFirstChainDot(SyntaxToken firstDot, Dictionary<SyntaxToken, SyntaxToken> replacements)
-    {
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstDot) == false
-            || HasIntermediateMemberAccess(firstDot))
-        {
-            return;
-        }
-        replacements[firstDot] = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(firstDot);
-
-        var previousToken = firstDot.GetPreviousToken();
-
-        if (previousToken != default
-            && previousToken.IsKind(SyntaxKind.None) == false
-            && LineBreakTriviaUtilities.HasTrailingEndOfLine(previousToken))
-        {
-            replacements[previousToken] = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia));
-        }
     }
 
     /// <summary>
@@ -885,7 +118,9 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="token">The token whose predecessor should be resolved</param>
     /// <param name="previousToken">Receives the previous token when one exists</param>
     /// <returns><see langword="true"/> if a previous token was found; otherwise, <see langword="false"/></returns>
-    private static bool TryGetPreviousToken<TNode>(TNode node, SyntaxToken token, out SyntaxToken previousToken)
+    protected static bool TryGetPreviousToken<TNode>(TNode node,
+                                                     SyntaxToken token,
+                                                     out SyntaxToken previousToken)
         where TNode : SyntaxNode
     {
         previousToken = default;
@@ -932,7 +167,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="node">The syntax node to inspect</param>
     /// <param name="token">The token to test</param>
     /// <returns><see langword="true"/> if the token lies within the node span; otherwise, <see langword="false"/></returns>
-    private static bool ContainsToken(SyntaxNode node, SyntaxToken token)
+    protected static bool ContainsToken(SyntaxNode node,
+                                        SyntaxToken token)
     {
         return token.IsKind(SyntaxKind.None) == false
                && token.FullSpan.Start >= node.FullSpan.Start
@@ -945,7 +181,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="node">The syntax node containing the token</param>
     /// <param name="token">The token to refresh</param>
     /// <returns>The current token from the node if it can be found; otherwise, the original token</returns>
-    private static SyntaxToken GetCurrentToken(SyntaxNode node, SyntaxToken token)
+    protected static SyntaxToken GetCurrentToken(SyntaxNode node,
+                                                 SyntaxToken token)
     {
         if (ContainsToken(node, token) == false)
         {
@@ -963,31 +200,6 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     }
 
     /// <summary>
-    /// Collapses a parameter list opener onto the same line as its declaration token
-    /// </summary>
-    /// <typeparam name="TNode">The syntax node type that owns the declaration</typeparam>
-    /// <param name="node">The syntax node that contains the declaration token and parameter list</param>
-    /// <param name="declarationToken">The declaration token that should share a line with the opening parenthesis</param>
-    /// <param name="parameterList">The parameter list to normalize</param>
-    /// <returns>The updated syntax node</returns>
-    private static TNode CollapseParameterListToDeclarationLine<TNode>(TNode node, SyntaxToken declarationToken, ParameterListSyntax parameterList)
-        where TNode : SyntaxNode
-    {
-        if (TokenGapUtilities.HasLineBreakBetween(declarationToken, parameterList.OpenParenToken) == false)
-        {
-            return node;
-        }
-
-        var newDeclarationToken = declarationToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingWhitespace(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(declarationToken.TrailingTrivia)));
-        var newOpenParen = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(parameterList.OpenParenToken);
-
-        return node.ReplaceTokens([declarationToken, parameterList.OpenParenToken],
-                                  (original, _) => original == declarationToken
-                                                       ? newDeclarationToken
-                                                       : newOpenParen);
-    }
-
-    /// <summary>
     /// Moves a token to a new line by prepending an end-of-line trivia to its leading trivia.
     /// Also strips any trailing whitespace from the previous token to avoid orphaned spaces
     /// </summary>
@@ -995,7 +207,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="node">The node containing the token</param>
     /// <param name="token">The token to move to a new line</param>
     /// <returns>The node with the token moved to a new line</returns>
-    private TNode MoveTokenToNewLine<TNode>(TNode node, SyntaxToken token)
+    protected TNode MoveTokenToNewLine<TNode>(TNode node,
+                                              SyntaxToken token)
         where TNode : SyntaxNode
     {
         var newToken = PrependEndOfLine(token);
@@ -1034,7 +247,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="token">The token whose leading trivia should be normalized</param>
     /// <param name="blankLineCount">The number of blank lines to preserve before the token</param>
     /// <returns>The updated token</returns>
-    private SyntaxToken NormalizeLeadingGap(SyntaxToken token, int blankLineCount)
+    protected SyntaxToken NormalizeLeadingGap(SyntaxToken token,
+                                              int blankLineCount)
     {
         var suffixStart = 0;
         var lastLeadingEndOfLineIndex = -1;
@@ -1077,7 +291,7 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
 
         for (var lineBreakIndex = 0; lineBreakIndex <= blankLineCount; lineBreakIndex++)
         {
-            newLeadingTrivia.Add(SyntaxFactory.EndOfLine(_context.EndOfLine));
+            newLeadingTrivia.Add(SyntaxFactory.EndOfLine(Context.EndOfLine));
         }
 
         newLeadingTrivia.AddRange(preservedLeadingTrivia);
@@ -1094,7 +308,10 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="withToken">Function that updates the token on the owning node</param>
     /// <param name="blankLineCount">The number of blank lines to preserve before the token</param>
     /// <returns>The updated node</returns>
-    private TNode NormalizeGapBeforeOwnedTokenPreservingPreviousTrivia<TNode>(TNode node, SyntaxToken token, Func<TNode, SyntaxToken, TNode> withToken, int blankLineCount)
+    protected TNode NormalizeGapBeforeOwnedTokenPreservingPreviousTrivia<TNode>(TNode node,
+                                                                                SyntaxToken token,
+                                                                                Func<TNode, SyntaxToken, TNode> withToken,
+                                                                                int blankLineCount)
         where TNode : SyntaxNode
     {
         token = GetCurrentToken(node, token);
@@ -1107,7 +324,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
         var hasPreviousToken = TryGetPreviousToken(node, token, out var previousToken);
         var hasLineBreak = hasPreviousToken && TokenGapUtilities.HasLineBreakBetween(previousToken, token);
         var currentBlankLineCount = hasPreviousToken
-                                        ? TokenGapUtilities.CountBlankLinesBetween(previousToken, token)
+                                        ? TokenGapUtilities.CountBlankLinesBetween(previousToken,
+                                                                                   token)
                                         : 0;
 
         if (hasLineBreak && currentBlankLineCount == blankLineCount)
@@ -1141,7 +359,10 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="withToken">Function that updates the token on the owning node</param>
     /// <param name="blankLineCount">The number of blank lines to preserve before the token</param>
     /// <returns>The updated node</returns>
-    private TNode NormalizeGapBeforeOwnedToken<TNode>(TNode node, SyntaxToken token, Func<TNode, SyntaxToken, TNode> withToken, int blankLineCount)
+    protected TNode NormalizeGapBeforeOwnedToken<TNode>(TNode node,
+                                                        SyntaxToken token,
+                                                        Func<TNode, SyntaxToken, TNode> withToken,
+                                                        int blankLineCount)
         where TNode : SyntaxNode
     {
         token = GetCurrentToken(node, token);
@@ -1160,7 +381,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
 
         var hasLineBreak = hasPreviousToken && TokenGapUtilities.HasLineBreakBetween(previousToken, token);
         var currentBlankLineCount = hasPreviousToken
-                                        ? TokenGapUtilities.CountBlankLinesBetween(previousToken, token)
+                                        ? TokenGapUtilities.CountBlankLinesBetween(previousToken,
+                                                                                   token)
                                         : 0;
 
         if (hasLineBreak && currentBlankLineCount == blankLineCount)
@@ -1191,7 +413,9 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="token">The token whose preceding gap should be normalized</param>
     /// <param name="blankLineCount">The number of blank lines to preserve before the token</param>
     /// <returns>The updated node</returns>
-    private TNode NormalizeGapBeforeToken<TNode>(TNode node, SyntaxToken token, int blankLineCount)
+    protected TNode NormalizeGapBeforeToken<TNode>(TNode node,
+                                                   SyntaxToken token,
+                                                   int blankLineCount)
         where TNode : SyntaxNode
     {
         token = GetCurrentToken(node, token);
@@ -1240,279 +464,6 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     }
 
     /// <summary>
-    /// Ensures sibling statements in a block start on separate lines
-    /// </summary>
-    /// <param name="node">The block node</param>
-    /// <returns>The updated block</returns>
-    private BlockSyntax EnsureStatementsStartOnSeparateLines(BlockSyntax node)
-    {
-        if (node.Statements.Count <= 1)
-        {
-            return node;
-        }
-
-        var modified = false;
-        var statements = node.Statements.ToArray();
-
-        for (var statementIndex = 1; statementIndex < statements.Length; statementIndex++)
-        {
-            if (statements[statementIndex - 1] is EmptyStatementSyntax
-                || statements[statementIndex] is EmptyStatementSyntax)
-            {
-                continue;
-            }
-
-            var previousToken = statements[statementIndex - 1].GetLastToken();
-            var currentToken = statements[statementIndex].GetFirstToken();
-
-            if (TokenGapUtilities.HasLineBreakBetween(previousToken, currentToken) == false)
-            {
-                statements[statementIndex] = NormalizeGapBeforeToken(statements[statementIndex], currentToken, blankLineCount: 0);
-                modified = true;
-
-                continue;
-            }
-
-            if (TokenGapUtilities.CountBlankLinesBetween(previousToken, currentToken) > 1)
-            {
-                statements[statementIndex] = NormalizeGapBeforeToken(statements[statementIndex], currentToken, blankLineCount: 1);
-                modified = true;
-            }
-        }
-
-        return modified
-                   ? node.WithStatements(SyntaxFactory.List(statements))
-                   : node;
-    }
-
-    /// <summary>
-    /// Ensures sibling statements in a switch section start on separate lines
-    /// </summary>
-    /// <param name="node">The switch section node</param>
-    /// <returns>The updated switch section</returns>
-    private SwitchSectionSyntax EnsureStatementsStartOnSeparateLines(SwitchSectionSyntax node)
-    {
-        if (node.Statements.Count <= 1)
-        {
-            return node;
-        }
-
-        var modified = false;
-        var statements = node.Statements.ToArray();
-
-        for (var statementIndex = 1; statementIndex < statements.Length; statementIndex++)
-        {
-            if (statements[statementIndex - 1] is EmptyStatementSyntax
-                || statements[statementIndex] is EmptyStatementSyntax)
-            {
-                continue;
-            }
-
-            var previousToken = statements[statementIndex - 1].GetLastToken();
-            var currentToken = statements[statementIndex].GetFirstToken();
-
-            if (TokenGapUtilities.HasLineBreakBetween(previousToken, currentToken) == false)
-            {
-                statements[statementIndex] = NormalizeGapBeforeToken(statements[statementIndex], currentToken, blankLineCount: 0);
-                modified = true;
-
-                continue;
-            }
-
-            if (TokenGapUtilities.CountBlankLinesBetween(previousToken, currentToken) > 1)
-            {
-                statements[statementIndex] = NormalizeGapBeforeToken(statements[statementIndex], currentToken, blankLineCount: 1);
-                modified = true;
-            }
-        }
-
-        return modified
-                   ? node.WithStatements(SyntaxFactory.List(statements))
-                   : node;
-    }
-
-    /// <summary>
-    /// Moves a trailing inline comment from an <c>if</c> condition onto its own line
-    /// </summary>
-    /// <param name="node">The if statement</param>
-    /// <returns>The updated if statement</returns>
-    private IfStatementSyntax MoveTrailingConditionCommentToOwnLine(IfStatementSyntax node)
-    {
-        var trailingTrivia = node.CloseParenToken.TrailingTrivia;
-        var commentIndex = -1;
-
-        for (var triviaIndex = 0; triviaIndex < trailingTrivia.Count; triviaIndex++)
-        {
-            if (trailingTrivia[triviaIndex].IsKind(SyntaxKind.SingleLineCommentTrivia)
-                || trailingTrivia[triviaIndex].IsKind(SyntaxKind.MultiLineCommentTrivia))
-            {
-                commentIndex = triviaIndex;
-
-                break;
-            }
-        }
-
-        if (commentIndex < 0)
-        {
-            return node;
-        }
-
-        for (var triviaIndex = commentIndex + 1; triviaIndex < trailingTrivia.Count; triviaIndex++)
-        {
-            if (trailingTrivia[triviaIndex].IsKind(SyntaxKind.WhitespaceTrivia)
-                || trailingTrivia[triviaIndex].IsKind(SyntaxKind.EndOfLineTrivia))
-            {
-                continue;
-            }
-
-            return node;
-        }
-
-        var commentTrivia = trailingTrivia[commentIndex];
-        var removeStart = commentIndex;
-
-        while (removeStart > 0 && trailingTrivia[removeStart - 1].IsKind(SyntaxKind.WhitespaceTrivia))
-        {
-            removeStart--;
-        }
-
-        var newTrailingTrivia = new List<SyntaxTrivia>(trailingTrivia.Count - (commentIndex - removeStart + 1));
-
-        for (var triviaIndex = 0; triviaIndex < removeStart; triviaIndex++)
-        {
-            newTrailingTrivia.Add(trailingTrivia[triviaIndex]);
-        }
-
-        for (var triviaIndex = commentIndex + 1; triviaIndex < trailingTrivia.Count; triviaIndex++)
-        {
-            newTrailingTrivia.Add(trailingTrivia[triviaIndex]);
-        }
-
-        var newLeadingTrivia = node.IfKeyword.LeadingTrivia.Add(commentTrivia)
-                                                           .Add(SyntaxFactory.EndOfLine(_context.EndOfLine));
-
-        return node.WithCloseParenToken(node.CloseParenToken.WithTrailingTrivia(SyntaxFactory.TriviaList(newTrailingTrivia)))
-                   .WithIfKeyword(node.IfKeyword.WithLeadingTrivia(newLeadingTrivia));
-    }
-
-    /// <summary>
-    /// Normalizes a method chain or conditional access chain.
-    /// For multi-line chains: collapses the first chain link to the root line
-    /// and ensures all subsequent links start on their own line.
-    /// Single-line chains are not modified
-    /// </summary>
-    /// <param name="node">The outermost chain node (invocation or conditional access)</param>
-    /// <returns>The node with normalized chain line breaks</returns>
-    private SyntaxNode NormalizeChain(SyntaxNode node)
-    {
-        var chainDots = new List<SyntaxToken>();
-
-        CollectChainLinkDots(node, chainDots);
-
-        if (chainDots.Count == 0)
-        {
-            return node;
-        }
-
-        if (chainDots.Count == 1)
-        {
-            return NormalizeSingleChainDot(node, chainDots[0]);
-        }
-
-        if (chainDots.Exists(LineBreakTriviaUtilities.HasLeadingEndOfLine) == false)
-        {
-            return node;
-        }
-
-        var replacements = new Dictionary<SyntaxToken, SyntaxToken>();
-
-        TryCollapseFirstChainDot(chainDots[0], replacements);
-        EnsureContinuationDotsStartOnNewLine(chainDots, replacements);
-
-        if (replacements.Count == 0)
-        {
-            return node;
-        }
-
-        return node.ReplaceTokens(replacements.Keys,
-                                  (original, _) => replacements[original]);
-    }
-
-    /// <summary>
-    /// Ensures continuation dots in a chain start on their own lines
-    /// </summary>
-    /// <param name="chainDots">The chain dot tokens</param>
-    /// <param name="replacements">The token replacement map to populate</param>
-    private void EnsureContinuationDotsStartOnNewLine(List<SyntaxToken> chainDots, Dictionary<SyntaxToken, SyntaxToken> replacements)
-    {
-        var endOfLine = SyntaxFactory.EndOfLine(_context.EndOfLine);
-
-        for (var dotIndex = 1; dotIndex < chainDots.Count; dotIndex++)
-        {
-            if (LineBreakTriviaUtilities.HasLeadingEndOfLine(chainDots[dotIndex]))
-            {
-                continue;
-            }
-
-            var newLeading = chainDots[dotIndex].LeadingTrivia.Insert(0, endOfLine);
-            replacements[chainDots[dotIndex]] = chainDots[dotIndex].WithLeadingTrivia(newLeading);
-
-            var previousToken = chainDots[dotIndex].GetPreviousToken();
-
-            if (previousToken != default
-                && previousToken.IsKind(SyntaxKind.None) == false
-                && replacements.ContainsKey(previousToken) == false
-                && previousToken.TrailingTrivia.Any(SyntaxKind.WhitespaceTrivia))
-            {
-                replacements[previousToken] = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingWhitespace(previousToken.TrailingTrivia));
-            }
-        }
-    }
-
-    /// <summary>
-    /// Ensures each expression in a collection or object initializer starts on its own line.
-    /// Also strips trailing whitespace from the previous token (typically a comma or open brace)
-    /// </summary>
-    /// <param name="node">The initializer expression node</param>
-    /// <returns>The initializer with each item on a separate line</returns>
-    private InitializerExpressionSyntax EnsureInitializerItemsOnSeparateLines(InitializerExpressionSyntax node)
-    {
-        for (var expressionIndex = 0; expressionIndex < node.Expressions.Count; expressionIndex++)
-        {
-            var expression = node.Expressions[expressionIndex];
-            var firstToken = expression.GetFirstToken();
-
-            if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstToken) == false)
-            {
-                node = MoveTokenToNewLine(node, firstToken);
-            }
-        }
-
-        return node;
-    }
-
-    /// <summary>
-    /// Ensures each member in an anonymous object creation expression starts on its own line
-    /// </summary>
-    /// <param name="node">The anonymous object creation expression node</param>
-    /// <returns>The node with each member on a separate line</returns>
-    private AnonymousObjectCreationExpressionSyntax EnsureAnonymousObjectMembersOnSeparateLines(AnonymousObjectCreationExpressionSyntax node)
-    {
-        for (var memberIndex = 0; memberIndex < node.Initializers.Count; memberIndex++)
-        {
-            var member = node.Initializers[memberIndex];
-            var firstToken = member.GetFirstToken();
-
-            if (LineBreakTriviaUtilities.HasLeadingEndOfLine(firstToken) == false)
-            {
-                node = MoveTokenToNewLine(node, firstToken);
-            }
-        }
-
-        return node;
-    }
-
-    /// <summary>
     /// Ensures an opening brace is on its own line by prepending an end-of-line trivia if missing.
     /// Also ensures the closing brace is on its own line
     /// </summary>
@@ -1523,7 +474,11 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="closeBrace">The close brace token</param>
     /// <param name="withCloseBrace">Function to replace the close brace on the node</param>
     /// <returns>The node with braces placed on their own lines</returns>
-    private TNode EnsureBraceOnOwnLine<TNode>(TNode node, SyntaxToken openBrace, Func<TNode, SyntaxToken, TNode> withOpenBrace, SyntaxToken closeBrace, Func<TNode, SyntaxToken, TNode> withCloseBrace)
+    protected TNode EnsureBraceOnOwnLine<TNode>(TNode node,
+                                                SyntaxToken openBrace,
+                                                Func<TNode, SyntaxToken, TNode> withOpenBrace,
+                                                SyntaxToken closeBrace,
+                                                Func<TNode, SyntaxToken, TNode> withCloseBrace)
         where TNode : SyntaxNode
     {
         if (closeBrace.IsMissing == false)
@@ -1547,7 +502,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="node">The node containing the opening brace</param>
     /// <param name="openBrace">The opening brace token</param>
     /// <returns>The node with the first content token on a new line</returns>
-    private TNode EnsureFirstContentOnNewLine<TNode>(TNode node, SyntaxToken openBrace)
+    protected TNode EnsureFirstContentOnNewLine<TNode>(TNode node,
+                                                       SyntaxToken openBrace)
         where TNode : SyntaxNode
     {
         if (openBrace.IsMissing)
@@ -1567,7 +523,6 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
             return node;
         }
 
-        // If the trailing trivia of the open brace already contains an end of line, no action needed
         if (LineBreakTriviaUtilities.HasTrailingEndOfLine(openBrace))
         {
             return node;
@@ -1583,7 +538,8 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// <param name="node">The node containing the closing brace</param>
     /// <param name="closeBrace">The closing brace token</param>
     /// <returns>The node with correct close-brace continuation</returns>
-    private TNode EnsureCloseBraceContinuation<TNode>(TNode node, SyntaxToken closeBrace)
+    protected TNode EnsureCloseBraceContinuation<TNode>(TNode node,
+                                                        SyntaxToken closeBrace)
         where TNode : SyntaxNode
     {
         if (closeBrace.IsMissing)
@@ -1598,7 +554,6 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
             return node;
         }
 
-        // These tokens are allowed on the same line after '}'
         if (nextToken.IsKind(SyntaxKind.SemicolonToken)
             || nextToken.IsKind(SyntaxKind.CommaToken)
             || nextToken.IsKind(SyntaxKind.CloseParenToken))
@@ -1617,268 +572,13 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     }
 
     /// <summary>
-    /// Normalizes binary operator position: if the operator is at the end of a line
-    /// (trailing trivia contains EndOfLine), moves it to the beginning of the next line.
-    /// This only moves the line break; indentation is handled separately
-    /// </summary>
-    /// <param name="node">The binary expression node</param>
-    /// <returns>The binary expression with the operator at the beginning of the continuation line</returns>
-    private BinaryExpressionSyntax NormalizeBinaryOperatorPosition(BinaryExpressionSyntax node)
-    {
-        var operatorToken = node.OperatorToken;
-
-        if (LineBreakTriviaUtilities.HasTrailingEndOfLine(operatorToken) == false)
-        {
-            return node;
-        }
-
-        // The operator is at the end of a line. Move the line break so the operator
-        // starts the next line instead.
-        // 1. Remove EndOfLine from operator trailing trivia
-        // 2. Append EndOfLine to left operand's last token trailing trivia
-        // 3. Clean leading EOL/whitespace from right operand's first token
-        var leftLastToken = node.Left.GetLastToken();
-
-        var newOperatorTrailing = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(operatorToken.TrailingTrivia);
-
-        var newLeftTrailing = AppendEndOfLine(leftLastToken.TrailingTrivia);
-
-        var newLeftLastToken = leftLastToken.WithTrailingTrivia(newLeftTrailing);
-        var newOperatorToken = operatorToken.WithTrailingTrivia(newOperatorTrailing);
-
-        var rightFirstToken = node.Right.GetFirstToken();
-        var newRightFirstToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(rightFirstToken);
-
-        node = node.ReplaceTokens([leftLastToken, operatorToken, rightFirstToken],
-                                  (original, _) =>
-                                  {
-                                      if (original == leftLastToken)
-                                      {
-                                          return newLeftLastToken;
-                                      }
-
-                                      if (original == operatorToken)
-                                      {
-                                          return newOperatorToken;
-                                      }
-
-                                      return newRightFirstToken;
-                                  });
-
-        return node;
-    }
-
-    /// <summary>
-    /// For multi-line conditional expressions, ensures <c>?</c> and <c>:</c> are placed on new lines.
-    /// If the operator is at the end of a line (trailing EndOfLine), it is moved to the beginning
-    /// of the next line instead
-    /// </summary>
-    /// <param name="node">The conditional expression node</param>
-    /// <returns>The conditional expression with ternary operators on new lines</returns>
-    private ConditionalExpressionSyntax NormalizeTernaryOperatorPosition(ConditionalExpressionSyntax node)
-    {
-        if (IsMultiLine(node) == false)
-        {
-            // Single-line nested ternaries must be broken to multi-line
-            if (node.WhenFalse is ConditionalExpressionSyntax == false
-                && node.WhenTrue is ConditionalExpressionSyntax == false)
-            {
-                return node;
-            }
-
-            var operatorTokens = new List<SyntaxToken>();
-
-            CollectTernaryOperatorTokens(node, operatorTokens);
-
-            node = node.ReplaceTokens(operatorTokens, (original, _) => PrependEndOfLine(original));
-        }
-
-        node = NormalizeQuestionTokenPosition(node);
-
-        return NormalizeColonTokenPosition(node);
-    }
-
-    /// <summary>
-    /// Normalizes placement of the <c>?</c> token in a ternary expression
-    /// </summary>
-    /// <param name="node">The conditional expression node</param>
-    /// <returns>The updated conditional expression</returns>
-    private ConditionalExpressionSyntax NormalizeQuestionTokenPosition(ConditionalExpressionSyntax node)
-    {
-        var questionToken = node.QuestionToken;
-
-        if (LineBreakTriviaUtilities.HasTrailingEndOfLine(questionToken))
-        {
-            return MoveQuestionTokenToNextLine(node, questionToken);
-        }
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(questionToken))
-        {
-            return node;
-        }
-
-        return node.WithQuestionToken(PrependEndOfLine(questionToken));
-    }
-
-    /// <summary>
-    /// Moves the ternary <c>?</c> token from line-end position to line-start position
-    /// </summary>
-    /// <param name="node">The conditional expression node</param>
-    /// <param name="questionToken">The question mark token</param>
-    /// <returns>The updated conditional expression</returns>
-    private ConditionalExpressionSyntax MoveQuestionTokenToNextLine(ConditionalExpressionSyntax node, SyntaxToken questionToken)
-    {
-        // ? is at end of condition line — move line break so ? starts the next line.
-        var conditionLastToken = node.Condition.GetLastToken();
-        var newQuestionTrailing = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(questionToken.TrailingTrivia);
-        var newConditionTrailing = AppendEndOfLine(conditionLastToken.TrailingTrivia);
-        var newConditionLastToken = conditionLastToken.WithTrailingTrivia(newConditionTrailing);
-        var newQuestionToken = questionToken.WithTrailingTrivia(newQuestionTrailing);
-
-        // Also collapse the true expression onto the same line as ?.
-        var whenTrueFirstToken = node.WhenTrue.GetFirstToken();
-
-        var newWhenTrueFirstToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(whenTrueFirstToken);
-
-        if (newWhenTrueFirstToken.LeadingTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-        {
-            newWhenTrueFirstToken = newWhenTrueFirstToken.WithLeadingTrivia(newWhenTrueFirstToken.LeadingTrivia.Add(SyntaxFactory.Space));
-        }
-
-        return node.ReplaceTokens([conditionLastToken, questionToken, whenTrueFirstToken],
-                                  (original, _) =>
-                                  {
-                                      if (original == conditionLastToken)
-                                      {
-                                          return newConditionLastToken;
-                                      }
-
-                                      if (original == questionToken)
-                                      {
-                                          return newQuestionToken;
-                                      }
-
-                                      return newWhenTrueFirstToken;
-                                  });
-    }
-
-    /// <summary>
-    /// Normalizes placement of the <c>:</c> token in a ternary expression
-    /// </summary>
-    /// <param name="node">The conditional expression node</param>
-    /// <returns>The updated conditional expression</returns>
-    private ConditionalExpressionSyntax NormalizeColonTokenPosition(ConditionalExpressionSyntax node)
-    {
-        var colonToken = node.ColonToken;
-
-        if (LineBreakTriviaUtilities.HasTrailingEndOfLine(colonToken))
-        {
-            // : is at end of line — move line break so : starts the next line.
-            var whenTrueLastToken = node.WhenTrue.GetLastToken();
-            var newColonTrailing = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(colonToken.TrailingTrivia);
-            var newWhenTrueTrailing = AppendEndOfLine(whenTrueLastToken.TrailingTrivia);
-            var newWhenTrueLastToken = whenTrueLastToken.WithTrailingTrivia(newWhenTrueTrailing);
-            var newColonToken = colonToken.WithTrailingTrivia(newColonTrailing);
-
-            return node.ReplaceTokens([whenTrueLastToken, colonToken],
-                                      (original, _) =>
-                                      {
-                                          if (original == whenTrueLastToken)
-                                          {
-                                              return newWhenTrueLastToken;
-                                          }
-
-                                          return newColonToken;
-                                      });
-        }
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(colonToken))
-        {
-            return node;
-        }
-
-        return node.WithColonToken(PrependEndOfLine(colonToken));
-    }
-
-    /// <summary>
-    /// Ensures the constructor initializer (<c>: base()</c> or <c>: this()</c>) starts on a new line
-    /// </summary>
-    /// <param name="node">The constructor declaration node</param>
-    /// <returns>The constructor declaration with the initializer on a new line</returns>
-    private ConstructorDeclarationSyntax EnsureConstructorInitializerOnNewLine(ConstructorDeclarationSyntax node)
-    {
-        if (node.Initializer == null)
-        {
-            return node;
-        }
-
-        var colonToken = node.Initializer.ColonToken;
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(colonToken))
-        {
-            return node;
-        }
-
-        var newColonToken = PrependEndOfLine(colonToken);
-
-        return node.WithInitializer(node.Initializer.WithColonToken(newColonToken));
-    }
-
-    /// <summary>
-    /// Ensures all <c>where</c> constraint clauses in a type declaration start on new lines
-    /// </summary>
-    /// <typeparam name="TNode">The type declaration syntax type</typeparam>
-    /// <param name="node">The type declaration node</param>
-    /// <returns>The node with <c>where</c> clauses on new lines</returns>
-    private TNode EnsureGenericConstraintsOnNewLines<TNode>(TNode node)
-        where TNode : SyntaxNode
-    {
-        var constraintClauses = GetConstraintClauses(node);
-
-        if (constraintClauses == null || constraintClauses.Value.Count == 0)
-        {
-            return node;
-        }
-
-        var modified = false;
-        var newClauses = new List<TypeParameterConstraintClauseSyntax>();
-
-        foreach (var clause in constraintClauses.Value)
-        {
-            var whereKeyword = clause.WhereKeyword;
-
-            if (LineBreakTriviaUtilities.HasLeadingEndOfLine(whereKeyword) == false)
-            {
-                var newWhereKeyword = PrependEndOfLine(whereKeyword);
-
-                newClauses.Add(clause.WithWhereKeyword(newWhereKeyword));
-
-                modified = true;
-            }
-            else
-            {
-                newClauses.Add(clause);
-            }
-        }
-
-        if (modified == false)
-        {
-            return node;
-        }
-
-        var newConstraintList = SyntaxFactory.List(newClauses);
-
-        return SetConstraintClauses(node, newConstraintList);
-    }
-
-    /// <summary>
     /// Prepends an end-of-line trivia to a token's leading trivia
     /// </summary>
     /// <param name="token">The token to modify</param>
     /// <returns>The token with an end-of-line trivia prepended to its leading trivia</returns>
-    private SyntaxToken PrependEndOfLine(SyntaxToken token)
+    protected SyntaxToken PrependEndOfLine(SyntaxToken token)
     {
-        var endOfLine = SyntaxFactory.EndOfLine(_context.EndOfLine);
+        var endOfLine = SyntaxFactory.EndOfLine(Context.EndOfLine);
         var newLeading = token.LeadingTrivia.Insert(0, endOfLine);
 
         return token.WithLeadingTrivia(newLeading);
@@ -1889,1245 +589,10 @@ internal sealed class LineBreakRewriter : CSharpSyntaxRewriter
     /// </summary>
     /// <param name="triviaList">The trivia list to extend</param>
     /// <returns>The trivia list with an end-of-line trivia appended</returns>
-    private SyntaxTriviaList AppendEndOfLine(SyntaxTriviaList triviaList)
+    protected SyntaxTriviaList AppendEndOfLine(SyntaxTriviaList triviaList)
     {
-        return triviaList.Add(SyntaxFactory.EndOfLine(_context.EndOfLine));
+        return triviaList.Add(SyntaxFactory.EndOfLine(Context.EndOfLine));
     }
 
     #endregion // Methods
-
-    #region CSharpSyntaxRewriter
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitBlock(BlockSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (BlockSyntax)base.VisitBlock(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-        node = EnsureStatementsStartOnSeparateLines(node);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitClassDeclaration(ClassDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ClassDeclarationSyntax)base.VisitClassDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureGenericConstraintsOnNewLines(node);
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitStructDeclaration(StructDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (StructDeclarationSyntax)base.VisitStructDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureGenericConstraintsOnNewLines(node);
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitInterfaceDeclaration(InterfaceDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (InterfaceDeclarationSyntax)base.VisitInterfaceDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureGenericConstraintsOnNewLines(node);
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitRecordDeclaration(RecordDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (RecordDeclarationSyntax)base.VisitRecordDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureGenericConstraintsOnNewLines(node);
-
-        if (node.OpenBraceToken.IsMissing == false)
-        {
-            node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-            node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-            node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitEnumDeclaration(EnumDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (EnumDeclarationSyntax)base.VisitEnumDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitNamespaceDeclaration(NamespaceDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (NamespaceDeclarationSyntax)base.VisitNamespaceDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitAccessorList(AccessorListSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (AccessorListSyntax)base.VisitAccessorList(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (IsAutoPropertyAccessorList(node))
-        {
-            return node;
-        }
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitSwitchStatement(SwitchStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (SwitchStatementSyntax)base.VisitSwitchStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitInitializerExpression(InitializerExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        var originalParent = node.Parent;
-
-        node = (InitializerExpressionSyntax)base.VisitInitializerExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        // Array initializers (new int[] { 1 }, new[] { 1 }) stay inline — no Allman placement
-        if (node.IsKind(SyntaxKind.ArrayInitializerExpression))
-        {
-            return node;
-        }
-
-        // For collection/object initializers with multiple items, ensure each item is on its own line
-        if (node.Expressions.Count > 1)
-        {
-            node = EnsureInitializerItemsOnSeparateLines(node);
-        }
-
-        // Collection initializers inside property assignments keep brace inline (e.g., Validators = { ... })
-        if (node.IsKind(SyntaxKind.CollectionInitializerExpression)
-            && originalParent is AssignmentExpressionSyntax)
-        {
-            node = node.WithOpenBraceToken(LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(node.OpenBraceToken));
-
-            if (node.CloseBraceToken.IsMissing == false && LineBreakTriviaUtilities.HasLeadingEndOfLine(node.CloseBraceToken) == false)
-            {
-                node = node.WithCloseBraceToken(PrependEndOfLine(node.CloseBraceToken));
-            }
-        }
-        else if (node != null)
-        {
-            node = EnsureBraceOnOwnLine(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), node.CloseBraceToken, (n, t) => n.WithCloseBraceToken(t));
-        }
-
-        if (node != null)
-        {
-            node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        }
-
-        if (node != null)
-        {
-            node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-        }
-
-        // Clean trailing whitespace before close brace when it was moved to a new line
-        if (node != null)
-        {
-            node = CleanupTrailingWhitespaceBeforeCloseBrace(node);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitAssignmentExpression(AssignmentExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (AssignmentExpressionSyntax)base.VisitAssignmentExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        // Collection initializers in property assignments keep brace inline — strip trailing EOL from '='
-        if (node.Right is InitializerExpressionSyntax init
-            && init.IsKind(SyntaxKind.CollectionInitializerExpression)
-            && LineBreakTriviaUtilities.HasTrailingEndOfLine(node.OperatorToken))
-        {
-            var newTrivia = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(node.OperatorToken.TrailingTrivia);
-
-            node = node.WithOperatorToken(node.OperatorToken.WithTrailingTrivia(newTrivia));
-        }
-
-        // Collection expressions in property assignments keep bracket on the '=' line
-        if (node.Right is CollectionExpressionSyntax)
-        {
-            var operatorToken = node.OperatorToken;
-            var openBracket = node.Right.GetFirstToken();
-
-            if (LineBreakTriviaUtilities.HasTrailingEndOfLine(operatorToken))
-            {
-                var newOperatorTrivia = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(operatorToken.TrailingTrivia);
-
-                if (newOperatorTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-                {
-                    newOperatorTrivia = newOperatorTrivia.Add(SyntaxFactory.Space);
-                }
-
-                var newOpenBracket = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(openBracket);
-
-                node = node.ReplaceTokens([operatorToken, openBracket],
-                                          (original, _) =>
-                                          {
-                                              if (original == operatorToken)
-                                              {
-                                                  return operatorToken.WithTrailingTrivia(newOperatorTrivia);
-                                              }
-
-                                              return newOpenBracket;
-                                          });
-            }
-        }
-
-        // Rules for simple assignment (=): enforce same-line placement of operator and value.
-        if (node.IsKind(SyntaxKind.SimpleAssignmentExpression))
-        {
-            // Rule 1: The equals operator must be on the same line as the assignment target.
-            var operatorToken = node.OperatorToken;
-
-            if (LineBreakTriviaUtilities.HasLeadingEndOfLine(operatorToken))
-            {
-                var newOperatorToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(operatorToken);
-
-                if (newOperatorToken.LeadingTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-                {
-                    newOperatorToken = newOperatorToken.WithLeadingTrivia(newOperatorToken.LeadingTrivia.Add(SyntaxFactory.Space));
-                }
-
-                var previousToken = operatorToken.GetPreviousToken();
-
-                if (previousToken != default
-                    && previousToken.IsKind(SyntaxKind.None) == false
-                    && LineBreakTriviaUtilities.HasTrailingEndOfLine(previousToken))
-                {
-                    var newPreviousToken = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia));
-
-                    node = node.ReplaceTokens([previousToken, operatorToken],
-                                              (original, _) =>
-                                              {
-                                                  if (original == previousToken)
-                                                  {
-                                                      return newPreviousToken;
-                                                  }
-
-                                                  return newOperatorToken;
-                                              });
-                }
-                else
-                {
-                    node = node.WithOperatorToken(newOperatorToken);
-                }
-            }
-
-            // Rule 2: The right-hand side must start on the same line as the equals operator.
-            // Collection expressions and collection initializers are already handled above.
-            if (node.Right is not CollectionExpressionSyntax
-                && node.Right is not InitializerExpressionSyntax
-                && LineBreakTriviaUtilities.HasTrailingEndOfLine(node.OperatorToken))
-            {
-                var rightFirstToken = node.Right.GetFirstToken();
-
-                if (rightFirstToken != default
-                    && LineBreakTriviaUtilities.HasLeadingEndOfLine(rightFirstToken))
-                {
-                    var currentOperatorToken = node.OperatorToken;
-                    var newOperatorTrivia = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(currentOperatorToken.TrailingTrivia);
-
-                    if (newOperatorTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-                    {
-                        newOperatorTrivia = newOperatorTrivia.Add(SyntaxFactory.Space);
-                    }
-
-                    var newRightFirstToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(rightFirstToken);
-
-                    node = node.ReplaceTokens([currentOperatorToken, rightFirstToken],
-                                              (original, _) =>
-                                              {
-                                                  if (original == currentOperatorToken)
-                                                  {
-                                                      return currentOperatorToken.WithTrailingTrivia(newOperatorTrivia);
-                                                  }
-
-                                                  return newRightFirstToken;
-                                              });
-                }
-            }
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitEqualsValueClause(EqualsValueClauseSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (EqualsValueClauseSyntax)base.VisitEqualsValueClause(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        // Collection expressions in field/variable declarations keep bracket on the '=' line
-        if (node.Value is CollectionExpressionSyntax)
-        {
-            var equalsToken = node.EqualsToken;
-            var openBracket = node.Value.GetFirstToken();
-
-            if (LineBreakTriviaUtilities.HasTrailingEndOfLine(equalsToken))
-            {
-                var newEqualsTrivia = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(equalsToken.TrailingTrivia);
-
-                if (newEqualsTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-                {
-                    newEqualsTrivia = newEqualsTrivia.Add(SyntaxFactory.Space);
-                }
-
-                var newOpenBracket = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(openBracket);
-
-                node = node.ReplaceTokens([equalsToken, openBracket],
-                                          (original, _) =>
-                                          {
-                                              if (original == equalsToken)
-                                              {
-                                                  return equalsToken.WithTrailingTrivia(newEqualsTrivia);
-                                              }
-
-                                              return newOpenBracket;
-                                          });
-            }
-        }
-
-        // Rule 2: The value must start on the same line as the equals operator.
-        // Collection expressions are already handled above.
-        if (node.Value is not CollectionExpressionSyntax
-            && LineBreakTriviaUtilities.HasTrailingEndOfLine(node.EqualsToken))
-        {
-            var equalsToken = node.EqualsToken;
-            var valueFirstToken = node.Value.GetFirstToken();
-
-            if (valueFirstToken != default
-                && LineBreakTriviaUtilities.HasLeadingEndOfLine(valueFirstToken))
-            {
-                var newEqualsTrivia = LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(equalsToken.TrailingTrivia);
-
-                if (newEqualsTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-                {
-                    newEqualsTrivia = newEqualsTrivia.Add(SyntaxFactory.Space);
-                }
-
-                var newValueFirstToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(valueFirstToken);
-
-                node = node.ReplaceTokens([equalsToken, valueFirstToken],
-                                          (original, _) =>
-                                          {
-                                              if (original == equalsToken)
-                                              {
-                                                  return equalsToken.WithTrailingTrivia(newEqualsTrivia);
-                                              }
-
-                                              return newValueFirstToken;
-                                          });
-            }
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitVariableDeclarator(VariableDeclaratorSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (VariableDeclaratorSyntax)base.VisitVariableDeclarator(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Initializer == null)
-        {
-            return node;
-        }
-
-        // Rule 1: The equals token must be on the same line as the variable/field name.
-        // Rule 2 (value on same line as =) is handled in VisitEqualsValueClause.
-        var equalsToken = node.Initializer.EqualsToken;
-
-        if (LineBreakTriviaUtilities.HasLeadingEndOfLine(equalsToken))
-        {
-            var newEqualsToken = LineBreakTriviaUtilities.RemoveLeadingEndOfLineAndWhitespace(equalsToken);
-
-            if (newEqualsToken.LeadingTrivia.Any(SyntaxKind.WhitespaceTrivia) == false)
-            {
-                newEqualsToken = newEqualsToken.WithLeadingTrivia(newEqualsToken.LeadingTrivia.Add(SyntaxFactory.Space));
-            }
-
-            var previousToken = equalsToken.GetPreviousToken();
-
-            if (previousToken != default
-                && previousToken.IsKind(SyntaxKind.None) == false
-                && LineBreakTriviaUtilities.HasTrailingEndOfLine(previousToken))
-            {
-                var newPreviousToken = previousToken.WithTrailingTrivia(LineBreakTriviaUtilities.RemoveTrailingEndOfLineTrivia(previousToken.TrailingTrivia));
-
-                node = node.ReplaceTokens([previousToken, equalsToken],
-                                          (original, _) =>
-                                          {
-                                              if (original == previousToken)
-                                              {
-                                                  return newPreviousToken;
-                                              }
-
-                                              return newEqualsToken;
-                                          });
-            }
-            else
-            {
-                node = node.ReplaceToken(equalsToken, newEqualsToken);
-            }
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitObjectCreationExpression(ObjectCreationExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ObjectCreationExpressionSyntax)base.VisitObjectCreationExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Initializer != null)
-        {
-            node = NormalizeGapBeforeOwnedTokenPreservingPreviousTrivia(node, node.Initializer.OpenBraceToken, (n, t) => n.ReplaceToken(node.Initializer.OpenBraceToken, t), blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.Initializer.OpenBraceToken);
-        }
-
-        return CleanupTrailingWhitespaceBeforeInitializerBrace(node, node.Initializer);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitImplicitObjectCreationExpression(ImplicitObjectCreationExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ImplicitObjectCreationExpressionSyntax)base.VisitImplicitObjectCreationExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Initializer != null)
-        {
-            node = NormalizeGapBeforeOwnedTokenPreservingPreviousTrivia(node, node.Initializer.OpenBraceToken, (n, t) => n.ReplaceToken(node.Initializer.OpenBraceToken, t), blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.Initializer.OpenBraceToken);
-        }
-
-        return CleanupTrailingWhitespaceBeforeInitializerBrace(node, node.Initializer);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitAnonymousObjectCreationExpression(AnonymousObjectCreationExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (AnonymousObjectCreationExpressionSyntax)base.VisitAnonymousObjectCreationExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        // For anonymous objects with multiple members, ensure each member is on its own line
-        if (node.Initializers.Count > 1)
-        {
-            node = EnsureAnonymousObjectMembersOnSeparateLines(node);
-        }
-
-        node = NormalizeGapBeforeOwnedTokenPreservingPreviousTrivia(node, node.OpenBraceToken, (n, t) => n.WithOpenBraceToken(t), blankLineCount: 0);
-        node = EnsureFirstContentOnNewLine(node, node.OpenBraceToken);
-        node = NormalizeGapBeforeToken(node, node.CloseBraceToken, blankLineCount: 0);
-        node = EnsureCloseBraceContinuation(node, node.CloseBraceToken);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitArgumentList(ArgumentListSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ArgumentListSyntax)base.VisitArgumentList(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseFirstArgumentToSameLine(node);
-
-        return EnsureArgumentsOnSeparateLines(node, _context.EndOfLine);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitBracketedArgumentList(BracketedArgumentListSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (BracketedArgumentListSyntax)base.VisitBracketedArgumentList(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseFirstBracketedArgumentToSameLine(node);
-
-        return EnsureBracketedArgumentsOnSeparateLines(node, _context.EndOfLine);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitAttributeArgumentList(AttributeArgumentListSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (AttributeArgumentListSyntax)base.VisitAttributeArgumentList(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseFirstAttributeArgumentToSameLine(node);
-
-        return EnsureAttributeArgumentsOnSeparateLines(node, _context.EndOfLine);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitParameterList(ParameterListSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ParameterListSyntax)base.VisitParameterList(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseOpenParenToDeclarationLine(node);
-        node = CollapseFirstParameterToSameLine(node);
-        node = CollapseSeparatorsToPreviousParameterLine(node, _context.EndOfLine);
-        node = CollapseCloseParenToParameterLine(node);
-
-        return EnsureParametersOnSeparateLines(node, _context.EndOfLine);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitSwitchSection(SwitchSectionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (SwitchSectionSyntax)base.VisitSwitchSection(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        node = EnsureStatementsStartOnSeparateLines(node);
-
-        if (node.Statements.Count > 0 && node.Statements[0] is BlockSyntax block)
-        {
-            node = NormalizeContainedBlock(node, block);
-        }
-
-        return node;
-    }
-
-    /// <summary>
-    /// Normalizes brace placement for a block contained by a parent syntax node
-    /// </summary>
-    /// <typeparam name="TNode">The parent syntax node type</typeparam>
-    /// <param name="node">The parent node that contains the block</param>
-    /// <param name="block">The contained block</param>
-    /// <returns>The updated parent node</returns>
-    private TNode NormalizeContainedBlock<TNode>(TNode node, BlockSyntax block)
-        where TNode : SyntaxNode
-    {
-        node = NormalizeGapBeforeOwnedTokenPreservingPreviousTrivia(node, block.OpenBraceToken, (n, t) => n.ReplaceToken(block.OpenBraceToken, t), blankLineCount: 0);
-        node = EnsureFirstContentOnNewLine(node, block.OpenBraceToken);
-        node = NormalizeGapBeforeToken(node, block.CloseBraceToken, blankLineCount: 0);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitForStatement(ForStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ForStatementSyntax)base.VisitForStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitForEachStatement(ForEachStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ForEachStatementSyntax)base.VisitForEachStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitForEachVariableStatement(ForEachVariableStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ForEachVariableStatementSyntax)base.VisitForEachVariableStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitWhileStatement(WhileStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (WhileStatementSyntax)base.VisitWhileStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitDoStatement(DoStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (DoStatementSyntax)base.VisitDoStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitUsingStatement(UsingStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (UsingStatementSyntax)base.VisitUsingStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitLockStatement(LockStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (LockStatementSyntax)base.VisitLockStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitFixedStatement(FixedStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (FixedStatementSyntax)base.VisitFixedStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitCheckedStatement(CheckedStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (CheckedStatementSyntax)base.VisitCheckedStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Block != null)
-        {
-            node = NormalizeContainedBlock(node, node.Block);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitTryStatement(TryStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (TryStatementSyntax)base.VisitTryStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        node = NormalizeContainedBlock(node, node.Block);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitCatchClause(CatchClauseSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (CatchClauseSyntax)base.VisitCatchClause(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        return NormalizeContainedBlock(node, node.Block);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitFinallyClause(FinallyClauseSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (FinallyClauseSyntax)base.VisitFinallyClause(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        return NormalizeContainedBlock(node, node.Block);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitAccessorDeclaration(AccessorDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (AccessorDeclarationSyntax)base.VisitAccessorDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Body != null)
-        {
-            node = NormalizeContainedBlock(node, node.Body);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitSimpleLambdaExpression(SimpleLambdaExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (SimpleLambdaExpressionSyntax)base.VisitSimpleLambdaExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Body is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitParenthesizedLambdaExpression(ParenthesizedLambdaExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ParenthesizedLambdaExpressionSyntax)base.VisitParenthesizedLambdaExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Body is BlockSyntax statementBlock)
-        {
-            node = NormalizeContainedBlock(node, statementBlock);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitAnonymousMethodExpression(AnonymousMethodExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (AnonymousMethodExpressionSyntax)base.VisitAnonymousMethodExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Block != null)
-        {
-            node = NormalizeContainedBlock(node, node.Block);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitBinaryExpression(BinaryExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (BinaryExpressionSyntax)base.VisitBinaryExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        return NormalizeBinaryOperatorPosition(node);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitConditionalExpression(ConditionalExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ConditionalExpressionSyntax)base.VisitConditionalExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        return NormalizeTernaryOperatorPosition(node);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitConstructorDeclaration(ConstructorDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (ConstructorDeclarationSyntax)base.VisitConstructorDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseParameterListToDeclarationLine(node, node.Identifier, node.ParameterList);
-
-        if (node.Initializer != null)
-        {
-            node = EnsureConstructorInitializerOnNewLine(node);
-        }
-
-        if (node.Body != null)
-        {
-            node = NormalizeGapBeforeToken(node, node.Body.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.Body.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, node.Body.CloseBraceToken, blankLineCount: 0);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitMethodDeclaration(MethodDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (MethodDeclarationSyntax)base.VisitMethodDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseParameterListToDeclarationLine(node, node.Identifier, node.ParameterList);
-        node = EnsureGenericConstraintsOnNewLines(node);
-
-        if (node.Body != null)
-        {
-            node = NormalizeGapBeforeToken(node, node.Body.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.Body.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, node.Body.CloseBraceToken, blankLineCount: 0);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitDelegateDeclaration(DelegateDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (DelegateDeclarationSyntax)base.VisitDelegateDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseParameterListToDeclarationLine(node, node.Identifier, node.ParameterList);
-        node = EnsureGenericConstraintsOnNewLines(node);
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitOperatorDeclaration(OperatorDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (OperatorDeclarationSyntax)base.VisitOperatorDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Body != null)
-        {
-            node = NormalizeGapBeforeToken(node, node.Body.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.Body.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, node.Body.CloseBraceToken, blankLineCount: 0);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitLocalFunctionStatement(LocalFunctionStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (LocalFunctionStatementSyntax)base.VisitLocalFunctionStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-        node = CollapseParameterListToDeclarationLine(node, node.Identifier, node.ParameterList);
-        node = EnsureGenericConstraintsOnNewLines(node);
-
-        if (node.Body != null)
-        {
-            node = NormalizeGapBeforeToken(node, node.Body.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.Body.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, node.Body.CloseBraceToken, blankLineCount: 0);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitPropertyDeclaration(PropertyDeclarationSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (PropertyDeclarationSyntax)base.VisitPropertyDeclaration(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.ExpressionBody != null)
-        {
-            node = CollapseExpressionBodiedProperty(node);
-        }
-
-        if (node.AccessorList != null && IsAutoPropertyAccessorList(node.AccessorList) == false)
-        {
-            node = NormalizeGapBeforeToken(node, node.AccessorList.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, node.AccessorList.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, node.AccessorList.CloseBraceToken, blankLineCount: 0);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitIfStatement(IfStatementSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        node = (IfStatementSyntax)base.VisitIfStatement(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (node.Statement is BlockSyntax statementBlock)
-        {
-            node = NormalizeGapBeforeToken(node, statementBlock.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, statementBlock.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, statementBlock.CloseBraceToken, blankLineCount: 0);
-        }
-
-        if (node.Else?.Statement is BlockSyntax elseBlock)
-        {
-            node = NormalizeGapBeforeToken(node, elseBlock.OpenBraceToken, blankLineCount: 0);
-            node = EnsureFirstContentOnNewLine(node, elseBlock.OpenBraceToken);
-            node = NormalizeGapBeforeToken(node, elseBlock.CloseBraceToken, blankLineCount: 0);
-        }
-
-        return MoveTrailingConditionCommentToOwnLine(node);
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitInvocationExpression(InvocationExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        var isOutermost = IsOutermostChainInvocation(node);
-
-        node = (InvocationExpressionSyntax)base.VisitInvocationExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (isOutermost)
-        {
-            return NormalizeChain(node);
-        }
-
-        return node;
-    }
-
-    /// <inheritdoc/>
-    public override SyntaxNode VisitConditionalAccessExpression(ConditionalAccessExpressionSyntax node)
-    {
-        _cancellationToken.ThrowIfCancellationRequested();
-
-        var isOutermost = node.Parent is not ConditionalAccessExpressionSyntax;
-
-        node = (ConditionalAccessExpressionSyntax)base.VisitConditionalAccessExpression(node);
-
-        if (node == null)
-        {
-            return null;
-        }
-
-        if (isOutermost && ContainsInvocation(node.WhenNotNull))
-        {
-            node = (ConditionalAccessExpressionSyntax)NormalizeChain(node);
-
-            node = CollapseMemberBindingToQuestionToken(node);
-        }
-
-        return node;
-    }
-
-    #endregion // CSharpSyntaxRewriter
 }
